requirements.txt
chromadb 
PyPDF2
sentence-transformers
transformers
openai
streamlit
python-dotenv
pydantic
colorama